Remove 'internal' module
