+++ src/fshelper/mod.rs
@@ -1,4 +1,6 @@
+use std::borrow::Cow;
 use std::env::current_dir;
+use std::ffi::OsStr;
 use std::fs;
 use std::io;
 #[cfg(any(unix, target_os = "redox"))]
@@ -62,5 +64,21 @@ pub fn is_empty(entry: &walk::DirEntry) -> bool {
         }
     } else {
         false
+    }
+}
+
+#[cfg(any(unix, target_os = "redox"))]
+pub fn osstr_to_bytes(input: &OsStr) -> Cow<[u8]> {
+    use std::os::unix::ffi::OsStrExt;
+    Cow::Borrowed(input.as_bytes())
+}
+
+#[cfg(windows)]
+pub fn osstr_to_bytes(input: &OsStr) -> Cow<[u8]> {
+    let string = input.to_string_lossy();
+
+    match string {
+        Cow::Owned(string) => Cow::Owned(string.into_bytes()),
+        Cow::Borrowed(string) => Cow::Borrowed(string.as_bytes()),
     }
 }
