--- src/internal/mod.rs
@@ -1,18 +1,0 @@
-use std::borrow::Cow;
-use std::ffi::OsStr;
-
-#[cfg(any(unix, target_os = "redox"))]
-pub fn osstr_to_bytes(input: &OsStr) -> Cow<[u8]> {
-    use std::os::unix::ffi::OsStrExt;
-    Cow::Borrowed(input.as_bytes())
-}
-
-#[cfg(windows)]
-pub fn osstr_to_bytes(input: &OsStr) -> Cow<[u8]> {
-    let string = input.to_string_lossy();
-
-    match string {
-        Cow::Owned(string) => Cow::Owned(string.into_bytes()),
-        Cow::Borrowed(string) => Cow::Borrowed(string.as_bytes()),
-    }
-}
